doc: Mentorship Program Teaching Lean
This Markdown document is a guide to my hands-on training program at the
Open-Source Contribution Academy (OSSCA), organized by NIPA's OSS
Support Center (Open UP).

I'll use "Theorem Proving in Lean 4" to teach Lean to my mentees for six
weeks this fall. We'll also create a git repository that contains
solutions to all the exercises in the text and translate the Natural
Number Game into Korean.