doc: Proposed Amendment to the Act, Version 2
I propose an amendment to the Act on the Protection of Children and
Youth Against Sex Offenses. This amendment will include altered or
synthetic materials that use photographs, videos, or the voice of a
minor in CYSEM (child or youth sexual exploitation material) when the
materials may cause sexual desire or shame against the will of the
minor.